requirements.txt
easyocr==1.7.2
fastapi==0.115.9
GitPython==3.1.44
google==3.0.0
google-generativeai==0.8.5
image==1.5.33
ninja==1.11.1.4
numpy==1.24.4
opencv-python-headless==4.11.0.86
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
sentence-transformers==4.1.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
tokenizers==0.21.1
torch==2.1.2
python-jose[cryptography]==3.3.0
asyncmy==0.2.10
uvicorn==0.34.2